Refactor manifest classes into separate files
